+++ website/members/forms.py
@@ -102,11 +102,19 @@ class CustomUserEditForm(wagtail.UserEditForm):
     """
     Custom form to edit users from within the wagtail admin interface.
     """
-    # TODO: Person number
-    birthday = forms.DateField(required=True, label=_('Birthday'))
-    person_number_ext = forms.CharField(
+    person_number = forms.CharField(
+        max_length=13,
+        min_length=12,
+        help_text=_('Person number using the YYYYMMDD-XXXX format.'),
+        validators=[validators.RegexValidator(
+            regex=r'^\d{8}\-?\d{4}$',
+            message=_('Use the format YYYYMMDD-XXXX for your person number.')
+        )],
+        widget=forms.TextInput(attrs={
+            'class': 'form-control',
+            'placeholder': 'YYYYMMDD-XXXX',
+        }),
         required=True,
-        label=_('Person number extension'),
     )
     phone_number = forms.CharField(
         required=False,
@@ -122,24 +130,48 @@ class CustomUserEditForm(wagtail.UserEditForm):
         label=_("Study Program"),
     )
     status = forms.ChoiceField(
-        required=False,
         choices=Member.MEMBERSHIP_CHOICES,
         label=_("Membership status"),
     )
 
     def __init__(self, *args, **kwargs):
-        super(CustomUserEditForm, self).__init__(*args, **kwargs)
+        instance = kwargs.get('instance', None)
+
+        initial = kwargs.pop('initial', {})
+        if instance is not None:
+            initial['person_number'] = instance.person_number()
+
+        super(CustomUserEditForm, self).__init__(
+            initial=initial, *args, **kwargs
+        )
+
+    def save(self, commit=True):
+        person_number = self.cleaned_data['person_number']
+        self.instance.birthday = datetime.strptime(
+            person_number[:8], "%Y%m%d"
+        ).date()
+        self.instance.person_number_ext = person_number[-4:]
+
+        return super(CustomUserEditForm, self).save(commit=commit)
 
 
 class CustomUserCreationForm(wagtail.UserCreationForm):
     """
     Custom form to create user from within the Wagtail admin user interface.
     """
-    # TODO: Person number
-    birthday = forms.DateField(required=True, label=_('Birthday'))
-    person_number_ext = forms.CharField(
+    person_number = forms.CharField(
+        max_length=13,
+        min_length=12,
+        help_text=_('Person number using the YYYYMMDD-XXXX format.'),
+        validators=[validators.RegexValidator(
+            regex=r'^\d{8}\-?\d{4}$',
+            message=_('Use the format YYYYMMDD-XXXX for your person number.')
+        )],
+        widget=forms.TextInput(attrs={
+            'class': 'form-control',
+            'placeholder': 'YYYYMMDD-XXXX',
+        }),
         required=True,
-        label=_('Person number extension'),
     )
     phone_number = forms.CharField(
         required=False,
@@ -161,4 +193,21 @@ class CustomUserCreationForm(wagtail.UserCreationForm):
     )
 
     def __init__(self, *args, **kwargs):
-        super(CustomUserCreationForm, self).__init__(*args, **kwargs)
+        instance = kwargs.get('instance', None)
+
+        initial = kwargs.pop('initial', {})
+        if instance is not None:
+            initial['person_number'] = instance.person_number()
+
+        super(CustomUserCreationForm, self).__init__(
+            initial=initial, *args, **kwargs
+        )
+
+    def save(self, commit=True):
+        person_number = self.cleaned_data['person_number']
+        self.instance.birthday = datetime.strptime(
+            person_number[:8], "%Y%m%d"
+        ).date()
+        self.instance.person_number_ext = person_number[-4:]
+
+        return super(CustomUserCreationForm, self).save(commit=commit)