+++ website/website/settings/base.py
@@ -202,5 +202,5 @@ WAGTAIL_FRONTEND_LOGIN_URL = LOGIN_URL
 
 WAGTAIL_USER_EDIT_FORM = 'members.forms.CustomUserEditForm'
 WAGTAIL_USER_CREATION_FORM = 'members.forms.CustomUserCreationForm'
-WAGTAIL_USER_CUSTOM_FIELDS = ['birthday', 'person_number_ext', 'phone_number',
-                              'registration_year', 'study', 'status']
+WAGTAIL_USER_CUSTOM_FIELDS = ['phone_number', 'registration_year', 'study',
+                              'status']